# docs/openapi/swagger.json
{
  "openapi" : "3.0.1",
  "info" : {
    "contact" : {
      "email" : "support@swagger.io",
      "name" : "API Support",
      "url" : "http://www.swagger.io/support"
    },
    "description" : "LocalPaas Backend",
    "license" : {
      "name" : "Apache 2.0",
      "url" : "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "termsOfService" : "http://swagger.io/terms/",
    "title" : "LocalPaas Backend",
    "version" : "0.1"
  },
  "servers" : [ {
    "url" : "/v1"
  } ],
  "paths" : {
...
    "/users/current/mfa/totp-begin-setup" : {
      "post" : {
        "description" : "Begins MFA TOTP authenticator setup",
        "operationId" : "beginMFATotpSetup",
        "requestBody" : {
          "content" : {
            "*/*" : {
              "schema" : {
                "$ref" : "#/components/schemas/github_com_localpaas_localpaas_localpaas_app_usecase_useruc_userdto.BeginMFATotpSetupReq"
              }
            }
          },
          "description" : "request data",
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/github_com_localpaas_localpaas_localpaas_app_usecase_useruc_userdto.BeginMFATotpSetupResp"
                }
              }
            },
            "description" : "OK"
          },
          "400" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/github_com_localpaas_localpaas_localpaas_app_apperrors.ErrorInfo"
                }
              }
            },
            "description" : "Bad Request"
          },
          "500" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/github_com_localpaas_localpaas_localpaas_app_apperrors.ErrorInfo"
                }
              }
            },
            "description" : "Internal Server Error"
          }
        },
        "summary" : "Begins MFA TOTP authenticator setup",
        "tags" : [ "users" ],
        "x-codegen-request-body-name" : "body"
      }
    },
    "/users/current/mfa/totp-complete-setup" : {
      "post" : {
        "description" : "Completes MFA TOTP authenticator setup",
        "operationId" : "completeMFATotpSetup",
        "requestBody" : {
          "content" : {
            "*/*" : {
              "schema" : {
                "$ref" : "#/components/schemas/github_com_localpaas_localpaas_localpaas_app_usecase_useruc_userdto.CompleteMFATotpSetupReq"
              }
            }
          },
          "description" : "request data",
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/github_com_localpaas_localpaas_localpaas_app_usecase_useruc_userdto.CompleteMFATotpSetupResp"
                }
              }
            },
            "description" : "OK"
          },
          "400" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/github_com_localpaas_localpaas_localpaas_app_apperrors.ErrorInfo"
                }
              }
            },
            "description" : "Bad Request"
          },
          "500" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/github_com_localpaas_localpaas_localpaas_app_apperrors.ErrorInfo"
                }
              }
            },
            "description" : "Internal Server Error"
          }
        },
        "summary" : "Completes MFA TOTP authenticator setup",
        "tags" : [ "users" ],
        "x-codegen-request-body-name" : "body"
      }
    },
    "/users/current/mfa/totp-remove" : {
      "post" : {
        "description" : "Removes MFA TOTP authenticator setup",
        "operationId" : "removeMFATotp",
        "requestBody" : {
          "content" : {
            "*/*" : {
              "schema" : {
                "$ref" : "#/components/schemas/github_com_localpaas_localpaas_localpaas_app_usecase_useruc_userdto.RemoveMFATotpReq"
              }
            }
          },
          "description" : "request data",
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/github_com_localpaas_localpaas_localpaas_app_usecase_useruc_userdto.RemoveMFATotpResp"
                }
              }
            },
            "description" : "OK"
          },
          "400" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/github_com_localpaas_localpaas_localpaas_app_apperrors.ErrorInfo"
                }
              }
            },
            "description" : "Bad Request"
          },
          "500" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/github_com_localpaas_localpaas_localpaas_app_apperrors.ErrorInfo"
                }
              }
            },
            "description" : "Internal Server Error"
          }
        },
        "summary" : "Removes MFA TOTP authenticator setup",
        "tags" : [ "users" ],
        "x-codegen-request-body-name" : "body"
      }
    },
    "/users/current/password" : {
      "patch" : {
        "description" : "Changes user password",
        "operationId" : "updateUserPassword",
        "requestBody" : {
          "content" : {
            "*/*" : {
              "schema" : {
                "$ref" : "#/components/schemas/github_com_localpaas_localpaas_localpaas_app_usecase_useruc_userdto.UpdatePasswordReq"
              }
            }
          },
          "description" : "request data",
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/github_com_localpaas_localpaas_localpaas_app_usecase_useruc_userdto.UpdatePasswordResp"
                }
              }
            },
            "description" : "OK"
          },
          "400" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/github_com_localpaas_localpaas_localpaas_app_apperrors.ErrorInfo"
                }
              }
            },
            "description" : "Bad Request"
          },
          "500" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/github_com_localpaas_localpaas_localpaas_app_apperrors.ErrorInfo"
                }
              }
            },
            "description" : "Internal Server Error"
          }
        },
        "summary" : "Changes user password",
        "tags" : [ "users" ],
        "x-codegen-request-body-name" : "body"
      }
    },
    "/users/invite" : {
      "post" : {
        "description" : "Invites a user to system",
        "operationId" : "inviteUser",
        "requestBody" : {
          "content" : {
            "*/*" : {
              "schema" : {
                "$ref" : "#/components/schemas/github_com_localpaas_localpaas_localpaas_app_usecase_useruc_userdto.InviteUserReq"
              }
            }
          },
          "description" : "request data",
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/github_com_localpaas_localpaas_localpaas_app_usecase_useruc_userdto.InviteUserResp"
                }
              }
            },
            "description" : "OK"
          },
          "400" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/github_com_localpaas_localpaas_localpaas_app_apperrors.ErrorInfo"
                }
              }
            },
            "description" : "Bad Request"
          },
          "500" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/github_com_localpaas_localpaas_localpaas_app_apperrors.ErrorInfo"
                }
              }
            },
            "description" : "Internal Server Error"
          }
        },
        "summary" : "Invites a user to system",
        "tags" : [ "users" ],
        "x-codegen-request-body-name" : "body"
      }
    },
    "/users/signup-begin" : {
      "post" : {
        "description" : "Begins user signup process",
        "operationId" : "beginUserSignup",
        "requestBody" : {
          "content" : {
            "*/*" : {
              "schema" : {
                "$ref" : "#/components/schemas/github_com_localpaas_localpaas_localpaas_app_usecase_useruc_userdto.BeginUserSignupReq"
              }
            }
          },
          "description" : "request data",
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/github_com_localpaas_localpaas_localpaas_app_usecase_useruc_userdto.BeginUserSignupResp"
                }
              }
            },
            "description" : "OK"
          },
          "400" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/github_com_localpaas_localpaas_localpaas_app_apperrors.ErrorInfo"
                }
              }
            },
            "description" : "Bad Request"
          },
          "500" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/github_com_localpaas_localpaas_localpaas_app_apperrors.ErrorInfo"
                }
              }
            },
            "description" : "Internal Server Error"
          }
        },
        "summary" : "Begins user signup process",
        "tags" : [ "users" ],
        "x-codegen-request-body-name" : "body"
      }
    },
    "/users/signup-complete" : {
      "post" : {
        "description" : "Completes user signup process",
        "operationId" : "completeUserSignup",
        "requestBody" : {
          "content" : {
            "*/*" : {
              "schema" : {
                "$ref" : "#/components/schemas/github_com_localpaas_localpaas_localpaas_app_usecase_useruc_userdto.CompleteUserSignupReq"
              }
            }
          },
          "description" : "request data",
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/github_com_localpaas_localpaas_localpaas_app_usecase_useruc_userdto.CompleteUserSignupResp"
                }
              }
            },
            "description" : "OK"
          },
          "400" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/github_com_localpaas_localpaas_localpaas_app_apperrors.ErrorInfo"
                }
              }
            },
            "description" : "Bad Request"
          },
          "500" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/github_com_localpaas_localpaas_localpaas_app_apperrors.ErrorInfo"
                }
              }
            },
            "description" : "Internal Server Error"
          }
        },
        "summary" : "Completes user signup process",
        "tags" : [ "users" ],
        "x-codegen-request-body-name" : "body"
      }
    },
...
  },
  "components" : {
    "schemas" : {
      "github_com_localpaas_localpaas_localpaas_app_apperrors.ErrorInfo" : {
        "properties" : {
          "cause" : {
            "type" : "string"
          },
          "code" : {
            "type" : "string"
          },
          "debugLog" : {
            "type" : "string"
          },
          "detail" : {
            "type" : "string"
          },
          "errors" : {
            "items" : {
              "$ref" : "#/components/schemas/github_com_localpaas_localpaas_localpaas_app_apperrors.InnerErrorInfo"
            },
            "type" : "array"
          },
          "stackTrace" : {
            "type" : "string"
          },
          "status" : {
            "type" : "integer"
          },
          "title" : {
            "type" : "string"
          },
          "type" : {
            "type" : "string"
          }
        },
        "required" : [ "code", "detail", "status", "title", "type" ],
        "type" : "object"
      },
      "github_com_localpaas_localpaas_localpaas_app_apperrors.InnerErrorInfo" : {
        "properties" : {
          "cause" : {
            "type" : "string"
          },
          "code" : {
            "type" : "string"
          },
          "message" : {
            "type" : "string"
          },
          "path" : {
            "type" : "string"
          }
        },
        "required" : [ "code", "message", "path" ],
        "type" : "object"
      },
...
      "github_com_localpaas_localpaas_localpaas_app_base.UserRole" : {
        "enum" : [ "admin", "member" ],
        "type" : "string",
        "x-enum-varnames" : [ "UserRoleAdmin", "UserRoleMember" ]
      },
      "github_com_localpaas_localpaas_localpaas_app_base.UserSecurityOption" : {
        "enum" : [ "enforce-sso", "password-2fa", "password-only" ],
        "type" : "string",
        "x-enum-varnames" : [ "UserSecurityEnforceSSO", "UserSecurityPassword2FA", "UserSecurityPasswordOnly" ]
      },
...
      "github_com_localpaas_localpaas_localpaas_app_basedto.BaseMeta" : {
        "properties" : {
          "code" : {
            "type" : "string"
          },
          "message" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "github_com_localpaas_localpaas_localpaas_app_basedto.ModuleAccessReq" : {
        "properties" : {
          "access" : {
            "$ref" : "#/components/schemas/github_com_localpaas_localpaas_localpaas_app_entity.AccessActions"
          },
          "id" : {
            "description" : "could be module name",
            "type" : "string"
          }
        },
        "required" : [ "access", "id" ],
        "type" : "object"
      },
      "github_com_localpaas_localpaas_localpaas_app_basedto.ObjectAccessReq" : {
        "properties" : {
          "access" : {
            "$ref" : "#/components/schemas/github_com_localpaas_localpaas_localpaas_app_entity.AccessActions"
          },
          "id" : {
            "type" : "string"
          }
        },
        "required" : [ "access", "id" ],
        "type" : "object"
      },
...
      "github_com_localpaas_localpaas_localpaas_app_entity.AccessActions" : {
        "properties" : {
          "delete" : {
            "type" : "boolean"
          },
          "read" : {
            "type" : "boolean"
          },
          "write" : {
            "type" : "boolean"
          }
        },
        "required" : [ "delete", "read", "write" ],
        "type" : "object"
      },
...
      "github_com_localpaas_localpaas_localpaas_app_usecase_useruc_userdto.BeginMFATotpSetupReq" : {
        "type" : "object"
      },
      "github_com_localpaas_localpaas_localpaas_app_usecase_useruc_userdto.BeginMFATotpSetupResp" : {
        "properties" : {
          "data" : {
            "$ref" : "#/components/schemas/github_com_localpaas_localpaas_localpaas_app_usecase_useruc_userdto.MFATotpSetupDataResp"
          },
          "meta" : {
            "$ref" : "#/components/schemas/github_com_localpaas_localpaas_localpaas_app_basedto.BaseMeta"
          }
        },
        "required" : [ "data", "meta" ],
        "type" : "object"
      },
      "github_com_localpaas_localpaas_localpaas_app_usecase_useruc_userdto.BeginUserSignupDataResp" : {
        "properties" : {
          "accessExpiration" : {
            "type" : "string"
          },
          "email" : {
            "type" : "string"
          },
          "mfaTotpSecret" : {
            "type" : "string"
          },
          "qrCode" : {
            "$ref" : "#/components/schemas/github_com_localpaas_localpaas_localpaas_app_usecase_useruc_userdto.MFATotpQRCodeResp"
          },
          "role" : {
            "$ref" : "#/components/schemas/github_com_localpaas_localpaas_localpaas_app_base.UserRole"
          },
          "securityOption" : {
            "$ref" : "#/components/schemas/github_com_localpaas_localpaas_localpaas_app_base.UserSecurityOption"
          }
        },
        "required" : [ "accessExpiration", "email", "role", "securityOption" ],
        "type" : "object"
      },
      "github_com_localpaas_localpaas_localpaas_app_usecase_useruc_userdto.BeginUserSignupReq" : {
        "properties" : {
          "inviteToken" : {
            "type" : "string"
          }
        },
        "required" : [ "inviteToken" ],
        "type" : "object"
      },
      "github_com_localpaas_localpaas_localpaas_app_usecase_useruc_userdto.BeginUserSignupResp" : {
        "properties" : {
          "data" : {
            "$ref" : "#/components/schemas/github_com_localpaas_localpaas_localpaas_app_usecase_useruc_userdto.BeginUserSignupDataResp"
          },
          "meta" : {
            "$ref" : "#/components/schemas/github_com_localpaas_localpaas_localpaas_app_basedto.BaseMeta"
          }
        },
        "required" : [ "data", "meta" ],
        "type" : "object"
      },
      "github_com_localpaas_localpaas_localpaas_app_usecase_useruc_userdto.CompleteMFATotpSetupReq" : {
        "properties" : {
          "passcode" : {
            "type" : "string"
          },
          "totpToken" : {
            "type" : "string"
          }
        },
        "required" : [ "passcode", "totpToken" ],
        "type" : "object"
      },
      "github_com_localpaas_localpaas_localpaas_app_usecase_useruc_userdto.CompleteMFATotpSetupResp" : {
        "properties" : {
          "meta" : {
            "$ref" : "#/components/schemas/github_com_localpaas_localpaas_localpaas_app_basedto.BaseMeta"
          }
        },
        "required" : [ "meta" ],
        "type" : "object"
      },
      "github_com_localpaas_localpaas_localpaas_app_usecase_useruc_userdto.CompleteUserSignupReq" : {
        "properties" : {
          "fullName" : {
            "type" : "string"
          },
          "inviteToken" : {
            "type" : "string"
          },
          "mfaTotpSecret" : {
            "type" : "string"
          },
          "passcode" : {
            "description" : "Required when security option is password-2fa",
            "type" : "string"
          },
          "password" : {
            "description" : "Required when security option is password-2fa/password-only",
            "type" : "string"
          },
          "photo" : {
            "$ref" : "#/components/schemas/github_com_localpaas_localpaas_localpaas_app_usecase_useruc_userdto.UserPhotoReq"
          }
        },
        "required" : [ "fullName", "inviteToken", "mfaTotpSecret", "passcode", "password", "photo" ],
        "type" : "object"
      },
      "github_com_localpaas_localpaas_localpaas_app_usecase_useruc_userdto.CompleteUserSignupResp" : {
        "properties" : {
          "meta" : {
            "$ref" : "#/components/schemas/github_com_localpaas_localpaas_localpaas_app_basedto.BaseMeta"
          }
        },
        "required" : [ "meta" ],
        "type" : "object"
      },
...
      "github_com_localpaas_localpaas_localpaas_app_usecase_useruc_userdto.InviteUserDataResp" : {
        "properties" : {
          "inviteLink" : {
            "type" : "string"
          }
        },
        "required" : [ "inviteLink" ],
        "type" : "object"
      },
      "github_com_localpaas_localpaas_localpaas_app_usecase_useruc_userdto.InviteUserReq" : {
        "properties" : {
          "accessExpiration" : {
            "type" : "string"
          },
          "email" : {
            "type" : "string"
          },
          "moduleAccesses" : {
            "items" : {
              "$ref" : "#/components/schemas/github_com_localpaas_localpaas_localpaas_app_basedto.ModuleAccessReq"
            },
            "type" : "array"
          },
          "projectAccesses" : {
            "items" : {
              "$ref" : "#/components/schemas/github_com_localpaas_localpaas_localpaas_app_basedto.ObjectAccessReq"
            },
            "type" : "array"
          },
          "role" : {
            "$ref" : "#/components/schemas/github_com_localpaas_localpaas_localpaas_app_base.UserRole"
          },
          "securityOption" : {
            "$ref" : "#/components/schemas/github_com_localpaas_localpaas_localpaas_app_base.UserSecurityOption"
          },
          "sendInviteEmail" : {
            "type" : "boolean"
          }
        },
        "required" : [ "accessExpiration", "email", "moduleAccesses", "projectAccesses", "role", "securityOption", "sendInviteEmail" ],
        "type" : "object"
      },
      "github_com_localpaas_localpaas_localpaas_app_usecase_useruc_userdto.InviteUserResp" : {
        "properties" : {
          "data" : {
            "$ref" : "#/components/schemas/github_com_localpaas_localpaas_localpaas_app_usecase_useruc_userdto.InviteUserDataResp"
          },
          "meta" : {
            "$ref" : "#/components/schemas/github_com_localpaas_localpaas_localpaas_app_basedto.BaseMeta"
          }
        },
        "required" : [ "data", "meta" ],
        "type" : "object"
      },
...
      "github_com_localpaas_localpaas_localpaas_app_usecase_useruc_userdto.MFATotpQRCodeResp" : {
        "properties" : {
          "dataBase64" : {
            "type" : "string"
          },
          "imageSize" : {
            "type" : "integer"
          },
          "imageType" : {
            "type" : "string"
          }
        },
        "required" : [ "dataBase64", "imageSize", "imageType" ],
        "type" : "object"
      },
      "github_com_localpaas_localpaas_localpaas_app_usecase_useruc_userdto.MFATotpSetupDataResp" : {
        "properties" : {
          "qrCode" : {
            "$ref" : "#/components/schemas/github_com_localpaas_localpaas_localpaas_app_usecase_useruc_userdto.MFATotpQRCodeResp"
          },
          "secret" : {
            "type" : "string"
          },
          "totpToken" : {
            "type" : "string"
          }
        },
        "required" : [ "qrCode", "secret", "totpToken" ],
        "type" : "object"
      },
      "github_com_localpaas_localpaas_localpaas_app_usecase_useruc_userdto.RemoveMFATotpReq" : {
        "properties" : {
          "passcode" : {
            "type" : "string"
          }
        },
        "required" : [ "passcode" ],
        "type" : "object"
      },
      "github_com_localpaas_localpaas_localpaas_app_usecase_useruc_userdto.RemoveMFATotpResp" : {
        "properties" : {
          "meta" : {
            "$ref" : "#/components/schemas/github_com_localpaas_localpaas_localpaas_app_basedto.BaseMeta"
          }
        },
        "required" : [ "meta" ],
        "type" : "object"
      },
      "github_com_localpaas_localpaas_localpaas_app_usecase_useruc_userdto.UpdatePasswordReq" : {
        "properties" : {
          "currentPassword" : {
            "type" : "string"
          },
          "newPassword" : {
            "type" : "string"
          }
        },
        "required" : [ "currentPassword", "newPassword" ],
        "type" : "object"
      },
      "github_com_localpaas_localpaas_localpaas_app_usecase_useruc_userdto.UpdatePasswordResp" : {
        "properties" : {
          "meta" : {
            "$ref" : "#/components/schemas/github_com_localpaas_localpaas_localpaas_app_basedto.BaseMeta"
          }
        },
        "required" : [ "meta" ],
        "type" : "object"
      },
...
      "github_com_localpaas_localpaas_localpaas_app_usecase_useruc_userdto.UserPhotoReq" : {
        "properties" : {
          "dataBase64" : {
            "type" : "string"
          },
          "fileName" : {
            "type" : "string"
          }
        },
        "required" : [ "dataBase64", "fileName" ],
        "type" : "object"
      },
...
    },
    "securitySchemes" : {
      "BasicAuth" : {
        "scheme" : "basic",
        "type" : "http"
      }
    }
  },
...
}